// web/src/http/schemas/webhook.ts
import { z } from "zod";

export const webhookListItemSchema = z.object({
    id: z.uuidv7(),
    method: z.string(),
    pathName: z.string(),
    createdAt: z.coerce.date(),
});

export const webhookListSchema = z.object({
    webhooks: z.array(webhookListItemSchema),
    nextCursor: z.string().nullable(),
});

export const webhookDetailSchema = z.object({
    id: z.uuidv7(),
    method: z.string(),
    pathName: z.string(),
    createdAt: z.coerce.date(),
    headers: z.record(z.string(), z.string()),
    body: z.string().nullable(),
    queryParams: z.record(z.string(), z.string()).nullable(),
    ip: z.string(),
    statusCode: z.number(),
    contentType: z.string().nullable(),
    contentLength: z.number().nullable(),
});